+++ ocd_backend/sources/ori.ibabs.yaml
@@ -56,7 +56,7 @@ _entities: &entities
     entity: reports
     doc_type: events
     extractor: ocd_backend.extractors.ibabs.IBabsReportsExtractor
-    transformer: ocd_backend.transformers.ibabs_meeting.report_item
+    transformer: ocd_backend.transformers.ibabs_report.report_item
     enrichers:
     - - ocd_backend.enrichers.media_enricher.static.gcs_static_media_enricher
       - # reserved for enricher_settings
@@ -146,7 +146,7 @@ ori.ibabs:
     - entity: reports
       doc_type: events
       extractor: ocd_backend.extractors.ibabs.IBabsReportsExtractor
-      transformer: ocd_backend.transformers.ibabs_meeting.report_item
+      transformer: ocd_backend.transformers.ibabs_report.report_item
       include: ".*"
       exclude: (moties|amendementen)
       pdf_max_pages: 0
